requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
h11==0.14.0
idna==3.4
outcome==1.2.0
psycopg2-binary==2.9.5
pycparser==2.21
PySocks==1.7.1
requests==2.28.2
selenium==4.8.2
selenium-stealth==1.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
